move streaming to a dedicated package. Fix exception wrapping for single
