+++ grpc/src/main/java/io/stargate/grpc/service/GrpcService.java
@@ -20,6 +20,10 @@ import io.grpc.stub.StreamObserver;
 import io.stargate.db.Persistence;
 import io.stargate.db.Persistence.Connection;
 import io.stargate.db.Result;
+import io.stargate.grpc.service.streaming.MessageStreamObserver;
+import io.stargate.grpc.service.streaming.StreamingBatchHandlerFactory;
+import io.stargate.grpc.service.streaming.StreamingExceptionHandler;
+import io.stargate.grpc.service.streaming.StreamingQueryHandlerFactory;
 import io.stargate.proto.QueryOuterClass.Batch;
 import io.stargate.proto.QueryOuterClass.Query;
 import io.stargate.proto.QueryOuterClass.Response;
