+++ testing/src/main/java/io/stargate/it/grpc/streaming/ExecuteQueryStreamingTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.stargate.it.grpc;
+package io.stargate.it.grpc.streaming;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -25,9 +25,8 @@ import io.stargate.grpc.Values;
 import io.stargate.it.driver.CqlSessionExtension;
 import io.stargate.it.driver.CqlSessionSpec;
 import io.stargate.it.driver.TestKeyspace;
-import io.stargate.proto.QueryOuterClass;
-import io.stargate.proto.QueryOuterClass.Batch;
-import io.stargate.proto.QueryOuterClass.Response;
+import io.stargate.it.grpc.GrpcIntegrationTest;
+import io.stargate.proto.QueryOuterClass.Query;
 import io.stargate.proto.QueryOuterClass.ResultSet;
 import io.stargate.proto.QueryOuterClass.StreamingResponse;
 import io.stargate.proto.StargateGrpc.StargateStub;
@@ -47,7 +46,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
     initQueries = {
       "CREATE TABLE IF NOT EXISTS test (k text, v int, PRIMARY KEY(k, v))",
     })
-public class ExecuteBatchStreamingTest extends GrpcIntegrationTest {
+public class ExecuteQueryStreamingTest extends GrpcIntegrationTest {
 
   @AfterEach
   public void cleanup(CqlSession session) {
@@ -55,7 +54,7 @@ public class ExecuteBatchStreamingTest extends GrpcIntegrationTest {
   }
 
   @Test
-  public void simpleStreamingBatch(@TestKeyspace CqlIdentifier keyspace) {
+  public void simpleStreamingQuery(@TestKeyspace CqlIdentifier keyspace) {
     List<StreamingResponse> responses = new CopyOnWriteArrayList<>();
 
     StargateStub stub = asyncStubWithCallCredentials();
@@ -73,44 +72,38 @@ public class ExecuteBatchStreamingTest extends GrpcIntegrationTest {
           public void onCompleted() {}
         };
 
-    StreamObserver<Batch> requestObserver = stub.executeBatchStream(responseStreamObserver);
+    StreamObserver<Query> requestObserver = stub.executeQueryStream(responseStreamObserver);
 
     requestObserver.onNext(
-        Batch.newBuilder()
-            .addQueries(cqlBatchQuery("INSERT INTO test (k, v) VALUES ('a', 1)"))
-            .addQueries(
-                cqlBatchQuery(
-                    "INSERT INTO test (k, v) VALUES (?, ?)", Values.of("b"), Values.of(2)))
-            .setParameters(batchParameters(keyspace))
-            .build());
+        cqlQuery("INSERT INTO test (k, v) VALUES ('a', 1)", queryParameters(keyspace)));
 
     requestObserver.onNext(
-        Batch.newBuilder()
-            .addQueries(
-                cqlBatchQuery(
-                    "INSERT INTO test (k, v) VALUES (?, ?)", Values.of("c"), Values.of(3)))
-            .setParameters(batchParameters(keyspace))
-            .build());
+        cqlQuery(
+            "INSERT INTO test (k, v) VALUES (?, ?)",
+            queryParameters(keyspace),
+            Values.of("b"),
+            Values.of(2)));
 
-    // all inserted records may be not visible to the subsequent SELECT query
-    // because all reactive calls are non-blocking. Therefore, we need to wait for response of two
-    // batch insert
+    // all inserted records may be not visible to the 3rd query (SELECT)
+    // because all calls are non-blocking. Therefore, we need to wait for response of two insert
     // queries
-    Awaitility.await().atMost(10, TimeUnit.SECONDS).until(() -> responses.size() == 2);
-    requestObserver.onCompleted();
+    Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> responses.size() == 2);
 
-    Response response =
-        stubWithCallCredentials()
-            .executeQuery(cqlQuery("SELECT * FROM test", queryParameters(keyspace)));
-    assertThat(response.hasResultSet()).isTrue();
-    ResultSet rs = response.getResultSet();
+    requestObserver.onNext(cqlQuery("SELECT * FROM test", queryParameters(keyspace)));
+    requestObserver.onCompleted();
+    Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> responses.size() == 3);
+
+    assertThat(responses.get(0)).isNotNull();
+    assertThat(responses.get(1)).isNotNull();
+    StreamingResponse response = responses.get(2);
+
+    assertThat(response.getResponse().hasResultSet()).isTrue();
+    ResultSet rs = response.getResponse().getResultSet();
     assertThat(new HashSet<>(rs.getRowsList()))
         .isEqualTo(
             new HashSet<>(
                 Arrays.asList(
-                    rowOf(Values.of("a"), Values.of(1)),
-                    rowOf(Values.of("b"), Values.of(2)),
-                    rowOf(Values.of("c"), Values.of(3)))));
+                    rowOf(Values.of("a"), Values.of(1)), rowOf(Values.of("b"), Values.of(2)))));
   }
 
   @Test
@@ -135,14 +128,10 @@ public class ExecuteBatchStreamingTest extends GrpcIntegrationTest {
           public void onCompleted() {}
         };
 
-    StreamObserver<QueryOuterClass.Batch> requestObserver =
-        stub.executeBatchStream(responseStreamObserver);
+    StreamObserver<Query> requestObserver = stub.executeQueryStream(responseStreamObserver);
 
     requestObserver.onNext(
-        Batch.newBuilder()
-            .addQueries(cqlBatchQuery("INSERT INTO not_existing (k, v) VALUES ('a', 1)"))
-            .setParameters(batchParameters(keyspace))
-            .build());
+        cqlQuery("INSERT INTO not_existing (k, v) VALUES ('a', 1)", queryParameters(keyspace)));
 
     Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> responses.size() == 1);
     requestObserver.onCompleted();
@@ -155,9 +144,9 @@ public class ExecuteBatchStreamingTest extends GrpcIntegrationTest {
   }
 
   @Test
-  public void streamingBatchQueryWithNextAndError(@TestKeyspace CqlIdentifier keyspace) {
-    List<StreamingResponse> responses = new CopyOnWriteArrayList<>();
+  public void streamingQueryWithNextAndError(@TestKeyspace CqlIdentifier keyspace) {
     AtomicReference<Throwable> error = new AtomicReference<>();
+    List<StreamingResponse> responses = new CopyOnWriteArrayList<>();
 
     StargateStub stub = asyncStubWithCallCredentials();
     StreamObserver<StreamingResponse> responseStreamObserver =
@@ -176,21 +165,15 @@ public class ExecuteBatchStreamingTest extends GrpcIntegrationTest {
           public void onCompleted() {}
         };
 
-    StreamObserver<Batch> requestObserver = stub.executeBatchStream(responseStreamObserver);
+    StreamObserver<Query> requestObserver = stub.executeQueryStream(responseStreamObserver);
 
     requestObserver.onNext(
-        Batch.newBuilder()
-            .addQueries(cqlBatchQuery("INSERT INTO test (k, v) VALUES ('a', 1)"))
-            .setParameters(batchParameters(keyspace))
-            .build());
+        cqlQuery("INSERT INTO test (k, v) VALUES ('a', 1)", queryParameters(keyspace)));
 
     requestObserver.onNext(
-        Batch.newBuilder()
-            .addQueries(cqlBatchQuery("INSERT INTO not_existing (k, v) VALUES ('a', 1)"))
-            .setParameters(batchParameters(keyspace))
-            .build());
+        cqlQuery("INSERT INTO not_existing (k, v) VALUES ('a', 1)", queryParameters(keyspace)));
 
-    Awaitility.await().atMost(10, TimeUnit.SECONDS).until(() -> responses.size() == 2);
+    Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> responses.size() == 2);
     requestObserver.onCompleted();
     assertThat(error.get()).isNull();
     assertThat(responses.get(0).getResponse()).isNotNull();
