+++ testing/src/main/java/io/stargate/it/grpc/streaming/ReactiveQueryTest.java
@@ -1,4 +1,4 @@
-package io.stargate.it.grpc;
+package io.stargate.it.grpc.streaming;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -10,6 +10,7 @@ import io.stargate.grpc.Values;
 import io.stargate.it.driver.CqlSessionExtension;
 import io.stargate.it.driver.CqlSessionSpec;
 import io.stargate.it.driver.TestKeyspace;
+import io.stargate.it.grpc.GrpcIntegrationTest;
 import io.stargate.proto.QueryOuterClass;
 import io.stargate.proto.ReactorStargateGrpc;
 import java.util.Arrays;
